Extract tests.
