+++ tests/filters_test.rs
@@ -69,50 +69,6 @@ fn can_split_by_column() {
 }
 
 #[test]
-fn str_can_only_apply_one() {
-    nu_error!(
-        output,
-        cwd("tests/fixtures/formats"),
-        "open caco3_plastics.csv | first 1 | str origin --downcase --upcase"
-    );
-
-    assert!(output.contains("Usage: str field [--downcase|--upcase|--to-int]"));
-}
-
-#[test]
-fn str_downcases() {
-    nu!(
-        output,
-        cwd("tests/fixtures/formats"),
-        "open caco3_plastics.csv | first 1 | str origin --downcase | get origin | echo $it"
-    );
-
-    assert_eq!(output, "spain");
-}
-
-#[test]
-fn str_upcases() {
-    nu!(
-        output,
-        cwd("tests/fixtures/formats"),
-        "open appveyor.yml | str environment.global.PROJECT_NAME --upcase | get environment.global.PROJECT_NAME | echo $it"
-    );
-
-    assert_eq!(output, "NUSHELL");
-}
-
-#[test]
-fn str_converts_to_int() {
-    nu!(
-        output,
-        cwd("tests/fixtures/formats"),
-        "open caco3_plastics.csv | first 1 | str tariff_item --to-int | where tariff_item == 2509000000 | get tariff_item | echo $it"
-    );
-
-    assert_eq!(output, "2509000000");
-}
-
-#[test]
 fn can_sum() {
     nu!(
         output,
@@ -122,6 +78,7 @@ fn can_sum() {
 
     assert_eq!(output, "203")
 }
+
 #[test]
 fn can_filter_by_unit_size_comparison() {
     nu!(
